Conferences are now static to ensure IDs are consistent between launches
